models lib is not installing DB
