+++ packages/models-library/tests/test_clusters.py
@@ -9,10 +9,8 @@ from models_library.clusters import (
     CLUSTER_USER_RIGHTS,
     DEFAULT_CLUSTER_ID,
     Cluster,
-    ClusterTypeInModel,
 )
 from pydantic import BaseModel, ValidationError
-from simcore_postgres_database.models.clusters import ClusterType
 
 
 @pytest.mark.parametrize(
@@ -90,23 +88,3 @@ def test_cluster_fails_when_owner_has_no_user_rights_if_default_cluster(
         modified_example["access_rights"][owner_gid] = CLUSTER_ADMIN_RIGHTS
         with pytest.raises(ValidationError):
             model_cls(**modified_example)
-
-
-def test_cluster_type_in_model_includes_postgres_database_model():
-    models_library_cluster_types_names: set[str] = {
-        t.name for t in set(ClusterTypeInModel)
-    }
-    postgres_library_cluster_types_names: set[str] = {t.name for t in set(ClusterType)}
-    assert postgres_library_cluster_types_names.issubset(
-        models_library_cluster_types_names
-    )
-
-    models_library_cluster_types_values: set[str] = {
-        t.value for t in set(ClusterTypeInModel)
-    }  # type: ignore
-    postgres_library_cluster_types_values: set[str] = {
-        t.value for t in set(ClusterType)
-    }
-    assert postgres_library_cluster_types_values.issubset(
-        models_library_cluster_types_values
-    )
